[tech] Switched to .jsx extensions and small refacto
