+++ client/app/components/app.jsx
@@ -1,8 +1,8 @@
 import { h, Component } from 'preact'
 
-import { I18n } from './plugins/preact-polyglot'
+import { I18n } from '../plugins/preact-polyglot'
 
-import Sidebar from './components/sidebar'
+import Sidebar from './sidebar'
 
 const App = ({ context, lang, categories, children }) => (
     <I18n context={context} lang={lang}>
