+++ client/app/components/connected_list.jsx
@@ -1,11 +1,11 @@
 import { h, render } from 'preact'
 import { translate } from '../plugins/preact-polyglot'
 
-const Discovery = ({ t }) => (
+const ConnectedList = ({ t }) => (
     <div class="content">
-        <h1>{t('my_accounts discovery title')}</h1>
+        <h1>{t('my_accounts connected title')}</h1>
         <article>{t('my_accounts coming_soon')}</article>
     </div>
 )
 
-export default translate()(Discovery)
+export default translate()(ConnectedList)
