--- client/app/components/lib/konnectors.js
@@ -1,8 +1,0 @@
-// Library related to konnectors
-
-// TODO a better way to grab all konnectors
-export default {
-  getKonnectors() {
-    return window.initKonnectors
-  }
-}
