+++ client/app/styles/index.styl
@@ -56,7 +56,7 @@ aside
         display: inline-block
         vertical-align: middle
 
-.konnectors-list
+.accounts-list
         display         flex
         flex-wrap       wrap
