+++ client/test/route.spec.js
@@ -1,6 +1,6 @@
 import { expect } from 'chai'
 import { h, render, rerender } from 'preact'
-import App from '../app/app'
+import App from '../app/components/app'
 
 // describe('App', () => {
 //     let scratch
